fix: clear PR 14 CI gates

+++ crates/videorc-backend/src/noise_cleanup.rs
@@ -1542,7 +1542,7 @@ fn available_space(path: &Path) -> Option<u64> {
         return None;
     }
     let stats = unsafe { stats.assume_init() };
-    Some((stats.f_bavail as u64).saturating_mul(stats.f_frsize))
+    Some(stats.f_bavail.saturating_mul(stats.f_frsize))
 }
 
 #[cfg(target_os = "windows")]
